source/requirements.txt
django==3.0.3
djangorestframework==3.11.0
psycopg2-binary==2.8.4
django-filter
ipdb
django-extensions
Werkzeug
# psycopg2==2.8.4